requirements.txt
alembic==1.0.3
python-dateutil==2.7.5
ipython==5.8
filemagic==1.6
freezegun==0.3.11
github3.py==1.2.0
gunicorn==19.7.1
# jira==2.0.0
psycopg2-binary==2.7.5
pycodestyle==2.4.0
py-trello==0.13.0
pyyaml==4.2b4
requests>=2.20.0
requests-oauthlib==1.0.0
requests-kerberos==0.12.0
simplejson==3.16.0
SQLAlchemy==1.2.14

### projectKeyOrID is not available as an argument for jira.boards yet as it has not been tagged in the release
### so using master from github instead
-e git://github.com/pycontribs/jira.git@master#egg=jira

Flask==1.0.2
